requirements.txt
astroid==2.4.2
colorama==0.4.4
haversine==2.3.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
pylint==2.6.0
rope==0.18.0
six==1.15.0
srt==3.4.1
toml==0.10.2
wrapt==1.12.1
